+++ spec/lib/karafka/params/params_spec.rb
@@ -56,11 +56,9 @@ RSpec.describe Karafka::Params::Params do
 
         it 'expect to build with additional values and value' do
           Timecop.freeze do
-            # We symbolize keys as hash with indifferent access will compare with
-            # a hash against string keys. This does not matter when accessing via
-            # string or symbols for single arguments
-            params = params_class.build(message, parser)
-            expect(params.symbolize_keys).to eq params_attributes
+            stringified = {}
+            params_class.build(message, parser).each { |k, v| stringified[k.to_sym] = v }
+            expect(stringified).to eq params_attributes
           end
         end
       end