+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - #354 - Expose consumer heartbeat
 - #377 - Remove the persistent setup in favor of persistence
 - #375 - Sidekiq Backend parser mismatch
+- #369 - Single consumer can support more than one topic
 
 ## 1.2.5
 - #354 - Expose consumer heartbeat